requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-cors==4.0.0
flasgger==0.9.7.1
minio==7.1.16
python-dotenv==1.0.0
strawberry-graphql[flask]==0.215.3
psycopg2-binary==2.9.7
alembic==1.12.1